handle when title is changing properly
